service: properly handle who argument on D-Bus kill calls

+++ src/mount.c
@@ -1777,11 +1777,12 @@ static int mount_kill(Unit *u, KillWho who, KillMode mode, int signo, DBusError 
                 return -ENOENT;
         }
 
-        if (m->control_pid > 0)
-                if (kill(m->control_pid, signo) < 0)
-                        r = -errno;
+        if (who == KILL_CONTROL || who == KILL_ALL)
+                if (m->control_pid > 0)
+                        if (kill(m->control_pid, signo) < 0)
+                                r = -errno;
 
-        if (mode == KILL_CONTROL_GROUP) {
+        if (who == KILL_ALL && mode == KILL_CONTROL_GROUP) {
                 int q;
 
                 if (!(pid_set = set_new(trivial_hash_func, trivial_compare_func)))
@@ -1795,7 +1796,7 @@ static int mount_kill(Unit *u, KillWho who, KillMode mode, int signo, DBusError 
                         }
 
                 if ((q = cgroup_bonding_kill_list(m->meta.cgroup_bondings, signo, false, pid_set)) < 0)
-                        if (r != -EAGAIN && r != -ESRCH && r != -ENOENT)
+                        if (q != -EAGAIN && q != -ESRCH && q != -ENOENT)
                                 r = q;
         }
